refactor(desktop): move the CSP into its own module
address pr-swarm finding: 75 lines of CSP logic inline in a 700-line
main.js, when desktop/ already follows a module-per-concern convention
(oauth.js, portInUse.js). Pure move; frontend dir is now an explicit
parameter instead of reaching for main.js module scope.
