+++ desktop/main.js
@@ -5,18 +5,17 @@ const {
 	ipcMain,
 	Menu,
 	nativeImage,
-	session,
 	shell,
 	Tray,
 	screen,
 } = require('electron')
 const path = require('path')
 const fs = require('fs')
-const crypto = require('crypto')
 const {execFile} = require('child_process')
 const express = require('express')
 const portInUse = require('./portInUse.js')
 const oauth = require('./oauth.js')
+const {applyContentSecurityPolicy} = require('./csp.js')
 
 const frontendPath = 'frontend/'
 const PROTOCOL = 'vikunja-desktop'
@@ -219,78 +218,6 @@ ipcMain.handle('oauth:refresh-token', async (_event, apiUrl, refreshToken) => {
 	return oauth.refreshAccessToken(apiUrl, refreshToken)
 })
 
-// ─── Content Security Policy ─────────────────────────────────────────
-// Matches <script> tags without a src attribute, i.e. those carrying inline code.
-const INLINE_SCRIPT_RE = /<script\b(?![^>]*\ssrc=)[^>]*>([\s\S]*?)<\/script>/gi
-
-// build.js rewrites the inline script that sets window.API_URL, so hash what's on disk.
-function inlineScriptHashes() {
-	const html = fs.readFileSync(path.join(__dirname, frontendPath, 'index.html'), 'utf8')
-	return [...html.matchAll(INLINE_SCRIPT_RE)]
-		.map(([, body]) => `'sha256-${crypto.createHash('sha256').update(body, 'utf8').digest('base64')}'`)
-}
-
-function buildContentSecurityPolicy() {
-	let scriptSrc = "'self'"
-	try {
-		const hashes = inlineScriptHashes()
-		if (hashes.length === 0) {
-			throw new Error('no inline script found in index.html')
-		}
-		scriptSrc += ' ' + hashes.join(' ')
-	} catch (err) {
-		// Blocking that script leaves window.API_URL undefined, which throws before the app
-		// mounts. A weaker policy beats a blank window.
-		console.warn('Could not hash inline scripts, falling back to script-src unsafe-inline:', err.message)
-		scriptSrc += " 'unsafe-inline'"
-	}
-
-	return [
-		"default-src 'self'",
-		`script-src ${scriptSrc}`,
-		// TipTap, vue3-notification and the emoji picker append <style> elements at runtime.
-		"style-src 'self' 'unsafe-inline'",
-		// Task descriptions embed remote images; data: for bundled icons, blob: for avatars,
-		// project backgrounds and the TOTP QR code.
-		'img-src * data: blob:',
-		"font-src 'self'",
-		// The API URL is user-configurable to any remote instance and the realtime socket URL
-		// is derived from it, so no origin allowlist is possible.
-		'connect-src * ws: wss:',
-		// PDF attachment previews render a blob: URL in an iframe.
-		'frame-src blob:',
-		// registerServiceWorker.ts registers sw.js.
-		"worker-src 'self'",
-		"object-src 'none'",
-		"base-uri 'self'",
-		"form-action 'self'",
-		"frame-ancestors 'none'",
-	].join('; ')
-}
-
-function applyContentSecurityPolicy(port) {
-	const csp = buildContentSecurityPolicy()
-
-	session.defaultSession.webRequest.onHeadersReceived(
-		{urls: [`http://127.0.0.1:${port}/*`]},
-		({responseHeaders, resourceType}, callback) => {
-			// Documents only: sw.js importScripts() the workbox runtime, and a service worker
-			// inherits the CSP of its own script response rather than the page's.
-			if (resourceType !== 'mainFrame') {
-				callback({})
-				return
-			}
-
-			const headers = Object.fromEntries(
-				Object.entries(responseHeaders ?? {})
-					.filter(([name]) => name.toLowerCase() !== 'content-security-policy'),
-			)
-			headers['Content-Security-Policy'] = [csp]
-			callback({responseHeaders: headers})
-		},
-	)
-}
-
 // ─── Express server ──────────────────────────────────────────────────
 function startServer(callback) {
 	const eApp = express()
@@ -310,7 +237,7 @@ function startServer(callback) {
 		const server = eApp.listen(port, '127.0.0.1', () => {
 			serverPort = server.address().port
 			console.log(`Server started on port ${serverPort}`)
-			applyContentSecurityPolicy(serverPort)
+			applyContentSecurityPolicy(serverPort, path.join(__dirname, frontendPath))
 			callback(serverPort)
 		})
 	})
